Nguyen hoang anh

+++ BE/src/main/java/com/example/demo/Service/Impl/SubMaterialServiceImpl.java
@@ -318,6 +318,10 @@ public class SubMaterialServiceImpl implements SubMaterialService {
                 employeeMaterials.setRequestProductsSubmaterials(requestProductsSubmaterials);
                 employeeMaterials.setEmployee(user);
 
+
+                // Lưu từng đối tượng và thêm vào danh sách kết quả
+                employeeMaterialsList.add(employeeMaterialRepository.save(employeeMaterials));
+
             }
             apiResponse.setResult(Collections.singletonList("Xuất đơn nguyên vật liệu cho đơn hàng thành công"));
             return ResponseEntity.ok(apiResponse);
@@ -389,34 +393,45 @@ public class SubMaterialServiceImpl implements SubMaterialService {
         Products products = productRepository.findById(product_id);
         List<ProductSubMaterials> list = productSubMaterialsRepository.findByProductID(product_id);
         List<ProductSubMaterials> productSubMaterialsList = new ArrayList<>();
-        if(!list.isEmpty()){
-            productSubMaterialsRepository.deleteAll(list);
-            for (Map.Entry<Integer, Double> entry : subMaterialQuantities.entrySet()) {
-                int subMaterialId = entry.getKey();
-                double quantity = entry.getValue();
-                SubMaterials subMaterial = subMaterialsRepository.findById1(subMaterialId);
-                ProductSubMaterials productSubMaterial = new ProductSubMaterials(subMaterial, products, quantity);
-                productSubMaterialsList.add(productSubMaterial);
+        List<Employeematerials> employeematerialsList = employeeMaterialRepository.findEmployeematerialsByProductId(product_id);
+        if (employeematerialsList != null) {
+            throw new AppException(ErrorCode.EMPLOYEE_MATERIAL_EXISTED);
+        } else {
+            if (!list.isEmpty()) {
+                for (Map.Entry<Integer, Double> entry : subMaterialQuantities.entrySet()) {
+                    int subMaterialId = entry.getKey();
+                    double quantity = entry.getValue();
+                    SubMaterials subMaterial = subMaterialsRepository.findById1(subMaterialId);
+                    ProductSubMaterials productSubMaterial = new ProductSubMaterials(subMaterial, products, quantity);
+                    productSubMaterialsList.add(productSubMaterial);
+                }
+                productSubMaterialsRepository.deleteAll(list);
+                productSubMaterialsRepository.saveAll(productSubMaterialsList);
             }
-            productSubMaterialsRepository.saveAll(productSubMaterialsList);
         }
         return productSubMaterialsList;
     }
+
     @Override
     public List<RequestProductsSubmaterials> EditSubMaterialRequestProduct(int request_product_id, Map<Integer, Double> subMaterialQuantities) {
         RequestProducts requestProducts = requestProductRepository.findById(request_product_id);
         List<RequestProductsSubmaterials> list = new ArrayList<>();
         List<RequestProductsSubmaterials> requestProductsSubmaterialsList = new ArrayList<>();
-        if(!list.isEmpty()) {
-            requestProductsSubmaterialsRepository.deleteAll(list);
-            for (Map.Entry<Integer, Double> entry : subMaterialQuantities.entrySet()) {
-                int subMaterialId = entry.getKey();
-                double quantity = entry.getValue();
-                SubMaterials subMaterial = subMaterialsRepository.findById1(subMaterialId);
-                RequestProductsSubmaterials requestProductsSubmaterials = new RequestProductsSubmaterials(subMaterial, requestProducts, quantity);
-                requestProductsSubmaterialsList.add(requestProductsSubmaterials);
+        List<Employeematerials> employeematerialsList = employeeMaterialRepository.findEmployeematerialsByRequestProductId(request_product_id);
+        if (employeematerialsList != null) {
+            throw new AppException(ErrorCode.EMPLOYEE_MATERIAL_EXISTED);
+        } else {
+            if(!list.isEmpty()) {
+                requestProductsSubmaterialsRepository.deleteAll(list);
+                for (Map.Entry<Integer, Double> entry : subMaterialQuantities.entrySet()) {
+                    int subMaterialId = entry.getKey();
+                    double quantity = entry.getValue();
+                    SubMaterials subMaterial = subMaterialsRepository.findById1(subMaterialId);
+                    RequestProductsSubmaterials requestProductsSubmaterials = new RequestProductsSubmaterials(subMaterial, requestProducts, quantity);
+                    requestProductsSubmaterialsList.add(requestProductsSubmaterials);
+                }
+                requestProductsSubmaterialsRepository.saveAll(requestProductsSubmaterialsList);
             }
-            requestProductsSubmaterialsRepository.saveAll(requestProductsSubmaterialsList);
         }
         return  requestProductsSubmaterialsList;
     }
